main/services/stt_npu.py
from __future__ import annotations
import numpy as np
from pathlib import Path
import yaml

# 假设用户将使用 ONNX Runtime + Vitis AI Execution Provider 来加速 Whisper
# 或者使用 AMD 推出的特定 RyzenAI 库
# 我们这里先实现一个通用的 ONNX Runtime 接口，可以指定 EP 为 VitisAI

try:
    import onnxruntime as ort
except ImportError:
    ort = None

class RyzenAIWhisperSTT:
    def __init__(
        self,
        model_path: str,
        provider: str = "VitisAIExecutionProvider",
        language: str = "zh",
    ) -> None:
        if ort is None:
            raise ImportError("Please install onnxruntime-vitisai to use RyzenAI NPU.")
        
        self.language = language
        # Vitis AI EP 驱动 AMD NPU
        # 注意：通常需要模型是针对 NPU 编译好的 .onnx + .json (config) 或者 .xmodel
        self.session = ort.InferenceSession(
            model_path,
            providers=[provider]
        )
        print(f"RyzenAI NPU STT initialized with provider: {provider}")

    def transcribe_wav(self, wav_path: Path) -> str:
        # 这里需要复杂的 ONNX Whisper 预处理逻辑 (Mel Spectrogram)
        # 为了简洁，通常会依赖配套的处理器或 transformers
        # 此处展示占位逻辑，实际中可能使用 amd-ryzenai-whisper 这类封装好的库
        print(f"Transcribing {wav_path} using AMD NPU...")
        
        # 实际推理流程（伪代码）：
        # 1. 加载音频并重采样至 16kHz
        # 2. 计算 Mel 频谱
        # 3. session.run(None, {"input_features": mel})
        # 4. 解码 tokens
        
        return "[RyzenAI NPU Transcription Placeholder]"

class STTFactory:
    @staticmethod
    def get_engine(config: dict):
        engine_type = config.get("engine", "faster-whisper")
        if engine_type == "whisper-ryzenai":
            return RyzenAIWhisperSTT(
                model_path=config.get("model_path"),
                provider=config.get("provider", "VitisAIExecutionProvider"),
                language=config.get("language", "zh")
            )
        else:
            from .stt import FasterWhisperSTT
            return FasterWhisperSTT(
                model_name=config.get("model_name", "tiny"),
                device=config.get("device", "cpu"),
                compute_type=config.get("compute_type", "int8"),
                language=config.get("language", "zh")
            )
