requirements-amd-npu.txt
onnxruntime-vitisai
onnx
numpy>=1.24.0
PyYAML>=6.0.0
sounddevice>=0.5.0
# whisper-ryzenai-specific-deps (if any)
